Update ProtDomRetriever.py
A better improvement, the previous change led to some errors. So I update choose_best_entry into choose_best_domain 

The code is now centered on domain and not entry, which means it will always return domain for each ENTRY and if these domains overlap on the sequence, we keep the longest domain. - longest identified domain rule ! 
Which means we can have for each accession hybrid results with two entries, let's say IPR000001 and SSF99999. It fundamentally does not change from the very first method, but this way, the algorithm is more comprehensive and more robust.